Correction de bugs de routes + verifier si film pas dans un panier quand on le supprime + modification de la gestion des commandes : valider que si en attente de validation, détail de la commande que si la commande pas encore validee

+++ src/AppBundle/Controller/AdminController.php
@@ -148,9 +148,20 @@ class AdminController extends Controller
     public function supprimerFilm(Film $film)
     {
         $em = $this->getDoctrine()->getManager();
-        $em->remove($film);
-        $em->flush();
+        $paniersRelieFilm = $em->getRepository('AppBundle:Panier')->findBy([
+            'filmId' => $film
+        ]);
+        if($paniersRelieFilm == null)
+        {
+            $erreur = false;
+            $em->remove($film);
+            $em->flush();
+        }
+        else
+        {
+            $erreur = "Ce film est present dans le panier d un utilisateur. Impossible de le supprimer.";
+        }
         $films = $em->getRepository('AppBundle:Film')->findAll();
-        return $this->render('admin/gestionFilms.html.twig', (['films' => $films]));
+        return $this->render('admin/gestionFilms.html.twig', (['films' => $films, 'erreur' => $erreur]));
     }
 }
